httpsrv/requirements.txt
flask==2.0.1
grpcio
grpcio-tools
requests
gevent
loguru